requirements.txt
Django
django-cms
django-filer
Pillow
